+++ packages/integration-boilerplate-node/src/service.ts
@@ -263,16 +263,17 @@ export class IntegrationService<
       const json = await response.json();
 
       if (response.status !== 200) {
-        throw new Error(JSON.stringify(json));
+        throw new GraphQLError(JSON.stringify(json), json);
       }
 
       if (json.errors) {
         const errors = json.errors.map((e: any) => e.message).join("\n");
-        throw new Error(errors);
+        throw new GraphQLError(errors, json);
       }
 
       return json;
     } catch (e) {
+      if (e instanceof GraphQLError) throw e;
       throw new GraphQLError((e as Error).message);
     }
   }