+++ packages/mint-components/src/components/sqm-portal-login/usePortalLogin.tsx
@@ -39,7 +39,10 @@ export function usePortalLogin(props) {
     }
   }, [errors]);
 
-  const errorMessage = errors?.response?.errors?.[0]?.message || error;
+  const errorMessage =
+    errors?.response?.errors?.[0]?.extensions?.message ||
+    errors?.response?.errors?.[0]?.message ||
+    error;
 
   return {
     states: {